change format for logger

+++ app/server/controllers/boardController.js
@@ -20,7 +20,7 @@ class BoardController{
     }
 
     async addBoard(req, res, next) {
-        if (!req.body || res.statusCode == "422") return next(new Error("Board is not added"));
+        if (res.statusCode == "422") return next(new Error("No correct data"));
 
         const { name, color, description, createAt } = req.body;
 
@@ -35,6 +35,8 @@ class BoardController{
     }
 
     async deleteBoardByName(req, res, next) {
+        if (res.statusCode == "422") return next(new Error("No correct data"));
+
         const boardName = req.params.name;
 
         const removeBoard = await this.service.deleteBoardByName({ name: boardName });
@@ -43,11 +45,13 @@ class BoardController{
     }
 
     async updateBoard(req, res, next) {
+        if (res.statusCode == "422") return next(new Error("No correct data"));
+
         const boardName = req.params.name;
         const newValues = req.body;
 
         const board = await this.service.updateBoard(boardName, newValues);
-        
+
         board ? res.send(board) : next(new Error("Board is not updated"));
     }
 }
